Copy service code from Launcher to Supervisor
We're copying the platform-specific code for shutting down services
from the Launcher to the Supervisor. The code will continue to exist
in the Launcher for the foreseeable future until we can reasonably
deprecate it.

In the meantime, the code that now exists in the Supervisor is the
same code, but modified to contend with the fact that we're now
killing a process that we ourselves didn't spawn.

Any additional code that deals with spawning these processes has been
removed from the Supervisor copy.

The existing `Process` struct abstraction is kept, for the time-being
at least. We can look at refactoring it to be more streamlined in the
future.

Signed-off-by: Christopher Maier <cmaier@chef.io>